+++ src/test_cpverifier/library-checker/hafnian_of_matrix.ds.test.cpp
@@ -11,7 +11,7 @@ int main() {
   mint::set_mod(998244353);
   std::ios::sync_with_stdio(false);
   std::cin.tie(nullptr);
-  u64 n;
+  u32 n;
   std::cin >> n;
   mat a(n, n);
   std::cin >> a;
